feat: New issue format with single rule table
Previously the rule table showed up in each comment which meant
maintainers had to scroll to the bottom of the issue to get the latest
status. With this change, the current table is always in the issue body,
and a trail of comments indicates the changes to that table and the score.

This change also removes the deprecated probot-scheduler package, and
removes state so it can scale horizontally.